+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-- Initial version.
+- Initial version of Mistral AI client for Dart
+- Supports Mistral AI API version 0.0.1